fixed dtype:

+++ gender_classifier_train_pitch_norm.py
@@ -260,7 +260,7 @@ def dataio_prepare(hparams):
 
         # sig = sb.dataio.dataio.read_audio(pitch_adjusted_file)
 
-        sig = torch.from_numpy(rec)
+        sig = torch.from_numpy(rec).float()
         return sig
 
     sb.dataio.dataset.add_dynamic_item(datasets, audio_pipeline)
